requirements_adapters.txt
accelerate==0.33.0
adapters==1.0.0
datasets==2.21.0
evaluate==0.4.2
huggingface-hub==0.24.5
Jinja2==3.1.3
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.2
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.1
tokenizers==0.19.1
torch==2.4.0+cu124
tqdm==4.66.5
transformers==4.43.4
tuned-lens==0.2.0
umap-learn==0.5.7
wandb==0.18.3
